Moved user's games to /dashboard
Resolves #176
