+++ imports/ui/dashboard.less
@@ -1,0 +1,21 @@
+.list-item {
+  .game-title {
+    font-size: 20px;
+    font-weight: bold;
+    font-style: italic;
+    color: #FF5555;
+  }
+
+  .game-players {
+    font-size: 15px;
+    font-style: italic;
+  }
+}
+
+.home-page-title {
+  display:inline-block;
+  width: 100%;
+  font-size: 40px;
+  font-weight: bold;
+  text-align: center;
+}
