--- imports/ui/home.html
@@ -1,48 +1,0 @@
-<template name="home">
-  <div style="align:center;">
-    <span class="home-page-title">Home</span>
-  </div>
-
-  <div id="gamelistTab" class="tab-pane fade in  active">
-    <ul class="list-group">
-      {{#each game in games}}
-        {{> home_game game=game}}
-      {{/each}}
-    </ul>
-  </div>
-</template>
-
-<template name="home_game">
-  <a href="/game/{{game._id}}" class="list-group-item">
-    <div class="list-item">
-      <table style="width:100%;table-layout:fixed;">
-        <tbody>
-          <tr>
-            <td>
-              <span class="game-title">{{title game}}</span>
-              {{#with gameState game._id}}
-                <span class="label {{label_style}}">{{state}}</span>
-              {{/with}}
-              <span class="badge">{{game.turn_number}}</span>
-            </td>
-            <td rowspan="2" width="60px" align="center">
-            
-            </td>
-            <td rowspan="2" width="20px" align="center">
-              {{#with game}}
-              <a href="#" class="deleteGame">
-                <span class="glyphicon glyphicon-remove"></span>
-              </a>
-              {{/with}}
-            </td>
-          </tr>
-          <tr>
-            <td>
-              <span class="game-players">{{game.creator.name}} vs. {{game.challenger.name}}</span>
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  </a>
-</template>
