+++ imports/ui/layout.html
@@ -14,9 +14,10 @@
       <div id="navbar" class="collapse navbar-collapse">
         <ul class="nav navbar-nav">
           <li><a href="/">Home</a></li>
-          <li><a href="/rules">Rules</a></li>
+          <li><a href="/dashboard">My Dashboard</a></li>
           <li><a href="/create">Create Game</a></li>
           <li><a href="/waiting">Waiting Room</a></li>
+          <li><a href="/rules">Rules</a></li>
         </ul>
         <ul class="nav navbar-nav navbar-right">
           <li>{{> atNavButton}}</li>
